refactor(xmtp_logging): split native/wasm into submodules
Addresses review feedback: instead of threading #[cfg(target_arch)] through
handle.rs and builder.rs, each is now a thin module that gates per-platform
submodules (handle/{native,wasm}, builder/{native,wasm}). The native handle
(file + telemetry slots, worker guards) and the browser handle (level filter
only) live in their own files, as do the two install() bodies.
